Extract tool-suggest wire helpers into codex-tools
