+++ codex-rs/tools/src/lib.rs
@@ -19,6 +19,7 @@ mod tool_config;
 mod tool_definition;
 mod tool_discovery;
 mod tool_spec;
+mod tool_suggest;
 mod utility_tool;
 mod view_image;
 
@@ -112,6 +113,13 @@ pub use tool_spec::create_image_generation_tool;
 pub use tool_spec::create_local_shell_tool;
 pub use tool_spec::create_tools_json_for_responses_api;
 pub use tool_spec::create_web_search_tool;
+pub use tool_suggest::TOOL_SUGGEST_APPROVAL_KIND_VALUE;
+pub use tool_suggest::ToolSuggestArgs;
+pub use tool_suggest::ToolSuggestMeta;
+pub use tool_suggest::ToolSuggestResult;
+pub use tool_suggest::all_suggested_connectors_picked_up;
+pub use tool_suggest::build_tool_suggestion_elicitation_request;
+pub use tool_suggest::verified_connector_suggestion_completed;
 pub use utility_tool::create_list_dir_tool;
 pub use utility_tool::create_test_sync_tool;
 pub use view_image::ViewImageToolOptions;
